pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "anyscrape"
version = "0.1.0"
description = "Autonomous multi-agent web scraping and research pipeline"
requires-python = ">=3.10"
dependencies = [
    "crawl4ai>=0.7.4",
    "ddgs>=6.1.0",
    "openai>=1.56.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
]

[project.scripts]
anyscrape = "anyscrape.cli:main"

